requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
dj-database-url==2.1.0
Django==4.2.5
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
